pyproject.toml
[project]
name = "swarm-agentes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "anthropic>=0.49.0",
  "openai>=1.72.0",
  "openai-agents>=0.0.9",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
]
